# docs/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "token": {
        "description": "For TCP. The Unix socket needs none: its file permissions are the access control.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "berth's API, served by `berth serve`. Each endpoint is one of berth's operations, with the rules the command line has: an operation that changes something needs a caller that may write; one that restarts a container needs a caller that may restart, and the org's name again as `confirm`. Secret values are never returned. Errors are berth.error/v1 documents.",
    "license": {
      "name": "MIT"
    },
    "title": "berth",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/orgs/{org}/logs/follow": {
      "get": {
        "description": "The log as server-sent events (berth.event/v1): one `output` event per line, until the client closes the connection.",
        "operationId": "org_logs_follow",
        "parameters": [
          {
            "in": "path",
            "name": "org",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The events."
          }
        },
        "summary": "Follow an org's container log",
        "x-berth-access": "read",
        "x-berth-operation": "logs"
      }
    },
...
  },
  "security": [
    {},
    {
      "token": []
    }
  ]
}